move registerLockedMouseEvents and registerHoveringMouseEvents to MouseController
