+++ Frontend/library/src/Inputs/InputClassesFactory.ts
@@ -6,9 +6,6 @@ import { MouseController } from './MouseController';
 import { TouchController } from './TouchController';
 import { GamePadController } from './GamepadController';
 import { Config, ControlSchemeType } from '../Config/Config';
-import { LockedMouseEvents } from './LockedMouseEvents';
-import { HoveringMouseEvents } from './HoveringMouseEvents';
-import { IMouseEvents } from './IMouseEvents';
 import { Logger } from '../Logger/Logger';
 import { CoordinateConverter } from '../Util/CoordinateConverter';
 import { StreamMessageController } from '../UeInstanceMessage/StreamMessageController';
@@ -61,108 +58,28 @@ export class InputClassesFactory {
         const mouseController = new MouseController(
             this.toStreamerMessagesProvider,
             this.videoElementProvider,
-            this.coordinateConverter
+            this.coordinateConverter,
+            this.activeKeys
         );
         mouseController.clearMouseEvents();
 
         switch (controlScheme) {
             case ControlSchemeType.LockedMouse:
-                this.registerLockedMouseEvents(mouseController);
+                mouseController.registerLockedMouseEvents(mouseController);
                 break;
             case ControlSchemeType.HoveringMouse:
-                this.registerHoveringMouseEvents(mouseController);
+                mouseController.registerHoveringMouseEvents(mouseController);
                 break;
             default:
                 Logger.Info(
                     Logger.GetStackTrace(),
                     'unknown Control Scheme Type Defaulting to Locked Mouse Events'
                 );
-                this.registerLockedMouseEvents(mouseController);
+                mouseController.registerLockedMouseEvents(mouseController);
                 break;
         }
 
         return mouseController;
-    }
-
-    /**
-     * Register a locked mouse class
-     * @param mouseController - a mouse controller instance
-     * @param playerStyleAttributesProvider - a player style attributes instance
-     */
-    registerLockedMouseEvents(mouseController: MouseController) {
-        const videoElementParent =
-            this.videoElementProvider.getVideoParentElement() as HTMLDivElement;
-        const lockedMouseEvents: IMouseEvents = new LockedMouseEvents(
-            this.videoElementProvider,
-            mouseController,
-            this.activeKeys
-        );
-
-        videoElementParent.requestPointerLock =
-            videoElementParent.requestPointerLock ||
-            videoElementParent.mozRequestPointerLock;
-        document.exitPointerLock =
-            document.exitPointerLock || document.mozExitPointerLock;
-
-        // minor hack to alleviate ios not supporting pointerlock
-        if (videoElementParent.requestPointerLock) {
-            videoElementParent.onclick = () => {
-                videoElementParent.requestPointerLock();
-            };
-        }
-
-        const lockStateChangeListener = () =>
-            lockedMouseEvents.lockStateChange();
-        document.addEventListener(
-            'pointerlockchange',
-            lockStateChangeListener,
-            false
-        );
-        document.addEventListener(
-            'mozpointerlockchange',
-            lockStateChangeListener,
-            false
-        );
-
-        videoElementParent.onmousedown = (mouseEvent: MouseEvent) =>
-            lockedMouseEvents.handleMouseDown(mouseEvent);
-        videoElementParent.onmouseup = (mouseEvent: MouseEvent) =>
-            lockedMouseEvents.handleMouseUp(mouseEvent);
-        videoElementParent.onwheel = (wheelEvent: WheelEvent) =>
-            lockedMouseEvents.handleMouseWheel(wheelEvent);
-        videoElementParent.ondblclick = (mouseEvent: MouseEvent) =>
-            lockedMouseEvents.handleMouseDouble(mouseEvent);
-        videoElementParent.pressMouseButtons = (mouseEvent: MouseEvent) =>
-            lockedMouseEvents.handlePressMouseButtons(mouseEvent);
-        videoElementParent.releaseMouseButtons = (mouseEvent: MouseEvent) =>
-            lockedMouseEvents.handleReleaseMouseButtons(mouseEvent);
-    }
-
-    /**
-     * Register a hovering mouse class
-     * @param mouseController - A mouse controller object
-     */
-    registerHoveringMouseEvents(mouseController: MouseController) {
-        const videoElementParent =
-            this.videoElementProvider.getVideoParentElement() as HTMLDivElement;
-        const hoveringMouseEvents = new HoveringMouseEvents(mouseController);
-
-        videoElementParent.onmousemove = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.updateMouseMovePosition(mouseEvent);
-        videoElementParent.onmousedown = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handleMouseDown(mouseEvent);
-        videoElementParent.onmouseup = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handleMouseUp(mouseEvent);
-        videoElementParent.oncontextmenu = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handleContextMenu(mouseEvent);
-        videoElementParent.onwheel = (wheelEvent: WheelEvent) =>
-            hoveringMouseEvents.handleMouseWheel(wheelEvent);
-        videoElementParent.ondblclick = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handleMouseDouble(mouseEvent);
-        videoElementParent.pressMouseButtons = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handlePressMouseButtons(mouseEvent);
-        videoElementParent.releaseMouseButtons = (mouseEvent: MouseEvent) =>
-            hoveringMouseEvents.handleReleaseMouseButtons(mouseEvent);
     }
 
     /**
